Added interface in StockedGood and FreshGood

+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/Insurable.java
@@ -1,5 +1,5 @@
 package kenakata.catalog;
 
 public interface Insurable {
-    int insure(int quantity);
+    long insure(long quantity);
 }
